Adjusted dockerfile to build jar automatically, added email validation and adjusted delete user response

+++ Backend/seasonal_job_matching/src/main/java/grad_project/seasonal_job_matching/controller/UserController.java
@@ -2,7 +2,6 @@ package grad_project.seasonal_job_matching.controller;
 
 import java.util.List;
 
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -16,6 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 import grad_project.seasonal_job_matching.dto.UserDTO;
 import grad_project.seasonal_job_matching.model.User;
 import grad_project.seasonal_job_matching.services.UserService;
+import jakarta.validation.Valid;
 
 
 
@@ -44,43 +44,40 @@ public class UserController {
             return ResponseEntity.ok("User found!");
         }
     }
-    // @GetMapping("/{email}")
-    // public Optional<User> findByEmail(@PathVariable String email){
-    //     return users_service.findUserbyEmail(email);
-    // }
     
     @PostMapping("/new")
-    public ResponseEntity<String> createUser(@RequestBody UserDTO userdto){//if user is from mobile than type is jobseeker, else it is employer  
+    public ResponseEntity<String> createUser(@Valid @RequestBody UserDTO userdto){//if user is from mobile than type is jobseeker, else it is employer  
         try {
             users_service.createUser(userdto);
             return ResponseEntity.ok("User created successfully");
         } catch (RuntimeException e) {
-            return ResponseEntity.status(HttpStatus.CONFLICT).body(e.getMessage());
+            return ResponseEntity.ok("Email already in use!");
         }
         
         
     }
 
     @PostMapping("/edit/{id}")
-    public ResponseEntity<String> editUser(@PathVariable long id, @RequestBody UserDTO dto){
+    public ResponseEntity<String> editUser(@PathVariable long id,@Valid @RequestBody UserDTO dto){
         try {
             users_service.editUser(dto, id);
-            return ResponseEntity.ok("User updated successfully!");
+            return ResponseEntity.ok("User updated successfully");
         } catch (RuntimeException e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
+            return ResponseEntity.ok("Update failed");
         }
     }
 
 
     @DeleteMapping("/delete/{id}")
-    public void deleteUser(@PathVariable Long id){
-        users_service.deleteUser(id);
+    public ResponseEntity<String> deleteUser(@PathVariable Long id){
+        if (ResponseEntity.ok()==findByID(id)) {
+            users_service.deleteUser(id);
+            return ResponseEntity.ok("User deleted successfully!");
+        }else{
+            return ResponseEntity.ok("User not found!");
+        }
+        
+
     }
-    //CRUD methods here, add 4 updates one for each field of jobseeker 
-    /*
-    Create- DONE
-    Read(All-DONE, single profile- DONE)
-    Update - NOT DONE
-    Delete - DONE
-     */
+
 }
